% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{LOAD.config}
\alias{LOAD.config}
\title{load configurations.
If a list, first load defaults, then overwrite with anything in the supplied config.
If filename, first load defaults then overwrite with anything in the txt file.
If 'default' then just load defaults.}
\usage{
LOAD.config(input_config)
}
\description{
load configurations.
If a list, first load defaults, then overwrite with anything in the supplied config.
If filename, first load defaults then overwrite with anything in the txt file.
If 'default' then just load defaults.
}
