% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{procOneP.Tablet}
\alias{procOneP.Tablet}
\title{procOne.Tablet
load one participant's tablet data, process, and generate figures}
\usage{
procOneP.Tablet(
  pid,
  tablet_dat,
  figdir = getwd(),
  resultsdir = getwd(),
  config = "default"
)
}
\arguments{
\item{pid}{Participant ID}

\item{tablet_dat}{can either be a list obtained with importRaw.Tablet, or a folder name to load}

\item{figdir}{where to save figures. Defaults to current directory. 'none' to skip saving figures}

\item{config}{can be 'default' to load default values, or a list with fields that overwrite the defaults, or (FEATURE NOT SUPPORTED YET) a text file containing field-value pairs to overwrite the defaults.}
}
\description{
procOne.Tablet
load one participant's tablet data, process, and generate figures
}
