% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{smtdemo}
\alias{smtdemo}
\title{smtdemo
Copy demo data to a target location, and
a demo script for loading and analyzing the data.}
\usage{
smtdemo(writedir = getwd())
}
\arguments{
\item{writedir}{Location to copy the demo data and script. Default = getwd()}
}
\description{
smtdemo
Copy demo data to a target location, and
a demo script for loading and analyzing the data.
}
