% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{importRaw.Tablet}
\alias{importRaw.Tablet}
\title{importRaw.Tablet}
\usage{
importRaw.Tablet(folder)
}
\arguments{
\item{folder}{the location of the folder containing the tablet accelerometer, touch, and event text files}
}
\value{
A list with subfields for each of the three data files
}
\description{
importRaw.Tablet
}
